altsec always handles clipboard
Originally, clipboard selections and properties were separated by clients
UIDs. Then more strict resource separation was introduced, and the
clipboard was sticked to the its client, that was not flexible. After that
a HandleClipboard xorg.conf option was introduced, that allowed focused
clients to read and write to the clipboard. Now, keep only this behavior
for clipboard cause other cases do not make much sence and to make the code
simpler.