altsec initial extensions handling
To make the extension zero- or as little configuration as possible, enlist
the extensions required for modern clients and graphical toolkits to work,
and allow only safe subset of their operations for untrusted clients.

Also remove XInputExtension workaround (that makes GTK+4 clients work),
because now we have more general solution.

Also fix the bug that does not construct allowed extensions list if
AllowedExts option is not set.