Account X11 clients
Add logic to account clients by their executable stats, root and user
namespace they belong. Thus we can treat them as a single entity in the case
their accounts are equal even if they are regustered as different
clients (and have a different PIDs), but still distinguish them if they
are chrooted or containerized.

This also makes Wine hack unnessecary.