Relax properties and selections
The properties and selections are used for inter-client communication,
restricting them to tough leads to breaking a lot of thing without
visible benefits. Let's instead protect conventionally known properties
and selections in the way that makes sense, and allow the rest as
relatively safe.

In this commit, the protected selections are PRIMARY and CLIPBOARD, the
rest are allowed in any way clients want.

In this commit, the properties are divided into three groups:

 * Window manager related properties;
 * Window related properties;
 * Other properties.

The window manager properties can be read by any clients, but can be
manipulated only by the window manager. The window properties can be
set only by client that owns the window, the rest are managed by the
following rules:

 * Any client can create any free non-protected property on any window,
   and after that became its owner.
 * The property owner can write data to that property.
 * The window owner to which the property is attached can read data
   from that property.
 * The window owner to which the property is attached can delete the
   property, and the property owner can do the same.

Trusted clients can manipulate any properties and selections.